Swing UI: Add Start/Stop Download menu items
This requires not using the "fireTableDataChanged()" JTable function
since it loses the selection and the menu items act on the currently
selected row. Instead, fireTableRowsInserted and fireTableRowsUpdated
are called when appropriate. To ease that(and maybe make a common
"AbstractTorrentModel" to share with the Qt GUI), the TorrentManager
is now in the TorrentTableModel class